test: speed up restore and trash test
